+++ bases/main/src/psi/launcher/extensions.clj
@@ -70,13 +70,15 @@
    {:psi/init 'psi.github.extension/init
     :source-policies
     {:development {:local/root "extensions/github"}
-     :installed   {:local/root "extensions/github"}}}
+     :installed   {:local/root "extensions/github"}
+     :jar         {:mvn/version :psi/release-version}}}
 
    'psi/edit-clj
    {:psi/init 'psi.edit-clj.extension/init
     :source-policies
     {:development {:local/root "extensions/edit-clj"}
-     :installed   {:local/root "extensions/edit-clj"}}}
+     :installed   {:local/root "extensions/edit-clj"}
+     :jar         {:mvn/version :psi/release-version}}}
 
    'psi/logprobs
    {:psi/init 'extensions.logprobs/init